function that prints all the elements of a listint_t list.

+++ 0x13-more_singly_linked_lists/0-print_listint.c
@@ -9,7 +9,7 @@
   */
 size_t print_listint(const listint_t *h)
 {
-	size_t k = 0;
+	unsigned int k = 0;
 
 	while (h)
 	{
